util: Change hash_table to use quadratic probing
This will allow us to remove the large static table and use a
power of two hash table size that we can compute on the fly.
We can use bitmasking instead of modulo to fit our hash in the table,
and it's less code.

By using the algorithm hash = sh + i/2 + i*i/2
we are guaranteed that all retries from the quad probing
are distinct, and so we should be able to completely fill the table.
This has been verified separately by Eric Anholt and Jason Ekstrand.
This passes the test added to exercise a worst case collision scenario.

This also exhibits much better collision avoidance performance than
the existing implementation; the worst case collision I've found by
running shader-db on a subset of my shader-collection is 78 runs of
the double hashing loop before insetion succeeded. With this
implementation the worst case is 25. With the existing implementation
99% of insetions had succeeded after 10 retries, while with the new
implementation that is dropped to 7. 99.5% of insertions succeeded after
13 retries, while with the new implementation that number is down to 8-9.
While the new implementation performs better on collisions in general,
the biggest difference is that the old one uses more retries to insert
an entry when it exhibits collisions. While the new implementation sees
12506 insertions that have not found a location after 10 retries,
that drops to only 74 after 20 retries. The old implementation sees
13864 insertions not having a spot after 16 retries. 1370 of these
have still not found a spot in the table after 26 retries.

V4: Feedback from Eric Anholt
   - Don't change load factor or starting size.

V3: Feedback from Connor Abbott
   - Remove hash_size table
   - Correct comment-style

   Feedback from Eric Anholt
   - Correct quadratic probing algorithm

   Feedback from Jason Ekstrand
   - Add "unreachable" if we fail to insert in table