util/tests: Expand collision test for hash table
Add a test to exercise a worst case collision scenario
that may cause us to not be able to find an empty
slot in the table even though it is not full.
This hits the bug in my last revision of the series
converting the hash table to quadratic probing.

V2: Feedback from Emil Velikov
    -Don't include code in the assert