nir: Add a value range propagation pass
V2:
   Add section to pessimize on branching on undef
   Incorporate some of Connors feedback
   Incorporate some of Matt's feedback
   Rework set_float_range API
   Reworked to not have a separate initialization run
   Handle loops better
   Squash a bug where we where adding ourself as user instead of the user

V3:
   Special case calculation of abs.
   Set the type for things we find are constant.
   We are now detecting bcsel(x, 0, 0) (four cases)

V4:
   Remove support for phi's:
   It was buggy and hard to get right, as type-information
   was hard to get hold of. A quick statistics gathering
   showed phi-nodes to amount to only 0.24% of the ssa-defs,
   and so I figured the potential gains would be negligible.
   With the newly added "check-users-what-type-we-are" and
   "check-our-sources-for-what-type-we-are" functions this
   decision can be revisited, as we now have a means of getting
   the type-information also for a phi-instruction.

V5:
   Add support for calculating a lot of operations if operands
   are not of such nature that they cause issues.

   We now detect the crazy pattern of:
	vec1 ssa_78 = fsat ssa_77		[0.0f,  1.0f]
	vec1 ssa_79 = fneg ssa_78		[-1.0f, 0.0f]
	vec1 ssa_80 = fadd ssa_79, 1.0f		[0.0f,  1.0f]
	vec1 ssa_81 = fmul ssa_80, ssa_80	[0.0f,  1.0f]
	vec1 ssa_84 = fmul ssa_81, ssa_81	[0.0f,  1.0f]
	vec1 ssa_87 = fmul ssa_84, 0.4f		[0.0f,  0.4f]
	vec1 ssa_88 = fadd ssa_87, 0.3f		[0.3f,  0.7f]
	vec1 ssa_232 = fmax ssa_88, 0.0f	<-- Useless

And also this more trivial one:
	vec1 ssa_58 = fsat ssa_57		[0.0f,  1.0f]
	vec1 ssa_59 = fmul ssa_58, 0.5f		[0.0f,  0.5f]
	vec1 ssa_60 = fadd ssa_59, 0.5f		[0.5f,  1.0f]
	vec1 ssa_61 = fmul ssa_60, ssa_60	[0.25f, 1.0f]
	vec1 ssa_62 = fmax ssa_61, 0.0f		<-- Useless

Shader-db run with INTEL_USE_NIR=1 compared to
master with INTEL_USE_NIR=1:

total instructions in shared programs: 1754178 -> 1754148 (-0.00%)
instructions in affected programs:     6013 -> 5983 (-0.50%)
helped:                                30
HURT:                                  0
GAINED:                                0
LOST:                                  0

From my collection of shaders these games are helped:
warsow, Dungeon Defenders and Brutal Legend.

We are now calculating the range for aproximately 4.5 % of
ssa-defs. This could be greatly increased if we reduce the
amount of cases where we set things to overdefined due to
not knowing better. Operations like fmul, fadd, fneg, bcsel,
fsat, flrp and fmax are supported. These are the top-6 alu
operations in a statistics-gathering I performed, and amount
to aproximately 90% of alu-ops on a scalar only run. If we
actually knew the input to these operations there is the
possibility that the effectiveness could be greatly improved.

A piglit-run shows no regression with or without INTEL_USE_NIR=1.

V6:
   Add a fix for type-check of source

V7:
   Adapt to changes since last rebase
   Linewrapping changes
   Adapt to relocation of compilers