nir: Add a loop analysis pass
The pass is now working! It detects the induction variables
and calculates the unroll count correctly on the one shader
I've tried it on as of now.

I've removed support for unsigned and float induction values
as for now, for the simple reason that they don't appear in
my shader-db collection, and so I don't see it as common enough
that we want to pollute the pass with this in the initial version.

I haven't looked over the coding style at all, so there is
probably a lot of style issues, so don't spend to much time
looking into that.

I also need to do something wrt swizzles.

V2: Rebase, adapt to removal of function overloads